# backend/openapi.json
{"openapi": "3.1.0", "info": {"title": "PrimeVPN Backend", "version": "0.1.0"}, "paths": {"/token": {"post": {"tags": ["Authentication"], "summary": "Login For Access Token", "operationId": "login_for_access_token_token_post", "requestBody": {"content": {"application/x-www-form-urlencoded": {"schema": {"$ref": "#/components/schemas/Body_login_for_access_token_token_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Token"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/users/": {"post": {"tags": ["Users"], "summary": "Create User", "operationId": "create_user_users__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserCreate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/users/me": {"get": {"tags": ["Users"], "summary": "Read User Me", "operationId": "read_user_me_users_me_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/servers/": {"get": {"tags": ["Servers"], "summary": "Read Servers", "operationId": "read_servers_servers__get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "skip", "in": "query", "required": false, "schema": {"type": "integer", "default": 0, "title": "Skip"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "default": 100, "title": "Limit"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/ServerResponse"}, "title": "Response Read Servers Servers  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["Servers"], "summary": "Create Server", "operationId": "create_server_servers__post", "security": [{"OAuth2PasswordBearer": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ServerCreate"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ServerResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/vpn/config/{server_id}": {"post": {"tags": ["VPN"], "summary": "Generate Config", "operationId": "generate_config_vpn_config__server_id__post", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "server_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Server Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/VPNConfigResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/admin/users": {"get": {"tags": ["Admin"], "summary": "Get All Users", "operationId": "get_all_users_admin_users_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/UserResponse"}, "type": "array", "title": "Response Get All Users Admin Users Get"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/admin/users/{user_id}/tier": {"patch": {"tags": ["Admin"], "summary": "Update User Tier", "operationId": "update_user_tier_admin_users__user_id__tier_patch", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "User Id"}}, {"name": "tier", "in": "query", "required": true, "schema": {"type": "string", "title": "Tier"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/admin/users/{user_id}/role": {"patch": {"tags": ["Admin"], "summary": "Update User Role", "operationId": "update_user_role_admin_users__user_id__role_patch", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "User Id"}}, {"name": "role", "in": "query", "required": true, "schema": {"type": "string", "title": "Role"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/admin/servers": {"get": {"tags": ["Admin"], "summary": "Get All Servers", "operationId": "get_all_servers_admin_servers_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ServerResponse"}, "type": "array", "title": "Response Get All Servers Admin Servers Get"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}, "post": {"tags": ["Admin"], "summary": "Create Server", "operationId": "create_server_admin_servers_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ServerCreate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ServerResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/admin/servers/{server_id}/toggle": {"patch": {"tags": ["Admin"], "summary": "Toggle Server", "operationId": "toggle_server_admin_servers__server_id__toggle_patch", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "server_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Server Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ServerResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/": {"get": {"summary": "Root", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"Body_login_for_access_token_token_post": {"properties": {"grant_type": {"anyOf": [{"type": "string", "pattern": "^password$"}, {"type": "null"}], "title": "Grant Type"}, "username": {"type": "string", "title": "Username"}, "password": {"type": "string", "format": "password", "title": "Password"}, "scope": {"type": "string", "title": "Scope", "default": ""}, "client_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Client Id"}, "client_secret": {"anyOf": [{"type": "string"}, {"type": "null"}], "format": "password", "title": "Client Secret"}}, "type": "object", "required": ["username", "password"], "title": "Body_login_for_access_token_token_post"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ServerCreate": {"properties": {"name": {"type": "string", "title": "Name"}, "location": {"type": "string", "title": "Location"}, "ip_address": {"type": "string", "title": "Ip Address"}, "vpn_subnet": {"type": "string", "title": "Vpn Subnet"}, "port": {"type": "integer", "title": "Port", "default": 51820}, "public_key": {"type": "string", "title": "Public Key"}, "is_premium": {"type": "boolean", "title": "Is Premium", "default": false}}, "type": "object", "required": ["name", "location", "ip_address", "vpn_subnet", "public_key"], "title": "ServerCreate"}, "ServerResponse": {"properties": {"name": {"type": "string", "title": "Name"}, "location": {"type": "string", "title": "Location"}, "ip_address": {"type": "string", "title": "Ip Address"}, "vpn_subnet": {"type": "string", "title": "Vpn Subnet"}, "port": {"type": "integer", "title": "Port", "default": 51820}, "public_key": {"type": "string", "title": "Public Key"}, "is_premium": {"type": "boolean", "title": "Is Premium", "default": false}, "id": {"type": "integer", "title": "Id"}, "is_active": {"type": "boolean", "title": "Is Active"}}, "type": "object", "required": ["name", "location", "ip_address", "vpn_subnet", "public_key", "id", "is_active"], "title": "ServerResponse"}, "Token": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "title": "Token Type"}}, "type": "object", "required": ["access_token", "token_type"], "title": "Token"}, "UserCreate": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "tier": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Tier", "default": "free"}, "role": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Role", "default": "user"}, "subscription_expires_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Subscription Expires At"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["email", "password"], "title": "UserCreate"}, "UserResponse": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "tier": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Tier", "default": "free"}, "role": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Role", "default": "user"}, "subscription_expires_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Subscription Expires At"}, "id": {"type": "integer", "title": "Id"}, "is_active": {"type": "boolean", "title": "Is Active"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["email", "id", "is_active", "created_at"], "title": "UserResponse"}, "VPNConfigResponse": {"properties": {"config": {"type": "string", "title": "Config"}, "private_key": {"type": "string", "title": "Private Key"}, "public_key": {"type": "string", "title": "Public Key"}, "server_ip": {"type": "string", "title": "Server Ip"}}, "type": "object", "required": ["config", "private_key", "public_key", "server_ip"], "title": "VPNConfigResponse"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"OAuth2PasswordBearer": {"type": "oauth2", "flows": {"password": {"scopes": {}, "tokenUrl": "token"}}}}}}